+++ baltip/baltip.tcl
@@ -9,7 +9,7 @@
 # License: MIT.
 # _______________________________________________________________________ #
 
-package provide baltip 1.0.1
+package provide baltip 1.0.2
 
 package require Tk
 
@@ -247,7 +247,7 @@ proc ::baltip::my::Show {w text force geo optvals} {
   if {$w ne "" && ![winfo exists $w]} return
   set win $w.w__BALTIP
   # keep the label's colors untouched (for apave package)
-  catch {::apave::paveObj untouchWidgets $win.label}
+  catch {::apave::obj untouchWidgets $win.label}
   set px [winfo pointerx .]
   set py [winfo pointery .]
   if {$geo ne ""} {                    ;# balloons not related to widgets