+++ baltip/pkgIndex.tcl
@@ -1,4 +1,4 @@
-package ifneeded baltip 1.0.1 [list source [file join $dir baltip.tcl]]
+package ifneeded baltip 1.0.2 [list source [file join $dir baltip.tcl]]
 
 namespace eval ::baltip {
 
@@ -14,8 +14,9 @@ The original code has been modified to make the tip:
   * be faded/destroyed after an interval defined by a caller
   * be enabled/disabled for all or specific widgets
   * be displayed at the screen's edges
-  * be displayed with given opacity
+  * be displayed under the host widget
   * be displayed as a stand-alone balloon message at given coordinates
+  * be displayed with given opacity, font, paddings, colors
   * have configure/cget etc. wrapped in Tcl ensemble for convenience
 
 The video introduction to *baltip* is presented by